client/requirements.txt
paho-mqtt==1.6.1
requests
# You may have to add more libraries
